requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn[standard]==0.23.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.5.2
python-dotenv==1.0.1
